requirements.txt
certifi==2018.11.29
chardet==3.0.4
dataclasses==0.6
defusedxml==0.5.0
Django==2.1.5
django-filter==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djongo==1.2.31
idna==2.8
Markdown==3.0.1
oauthlib==3.0.1
pkg-resources==0.0.0
pymongo==3.7.2
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.24.1
whitenoise==4.1.2
